Move setup helper function `loginWithSaml` to SetupHelpers.
